create and move the items in the categories app
