+++ opentech/apply/categories/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CategoriesConfig(AppConfig):
+    name = 'categories'
