+++ opentech/apply/funds/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 from modelcluster.fields import ParentalKey
-from modelcluster.models import ClusterableModel
 from wagtail.wagtailadmin.edit_handlers import (
     FieldPanel,
     InlinePanel,
@@ -20,6 +19,7 @@ WORKFLOW_CLASS = {
     SingleStage.name: SingleStage,
     DoubleStage.name: DoubleStage,
 }
+
 
 class FundType(AbstractStreamForm):
     parent_page_types = ['apply_home.ApplyHomePage']
@@ -67,33 +67,3 @@ class ApplicationForm(models.Model):
 
     def __str__(self):
         return self.name
-
-
-class Option(Orderable):
-    value = models.CharField(max_length=255)
-    category = ParentalKey('Category', related_name='options')
-
-
-class Category(ClusterableModel):
-    """Used to manage the global select questions used in most of the application form
-    Also used in the front end by editors when writing about projects.
-
-    When used in a form: name -> field label and help_text -> help_text
-    """
-    name = models.CharField(max_length=255)
-    help_text = models.CharField(max_length=255, blank=True)
-
-    @property
-    def field_label(self):
-        return self.name
-
-    panels = [
-        FieldPanel('name'),
-        InlinePanel('options', label='Options'),
-    ]
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name_plural = 'Categories'
